chore: noop generic handler and tests
