+++ test/digest/handler-noop.test.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Adobe. All rights reserved.
+ * Copyright 2024 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -11,32 +11,47 @@
  */
 /* eslint-env mocha */
 
+import sinon from 'sinon';
 import chai from 'chai';
+import sinonChai from 'sinon-chai';
 import chaiAsPromised from 'chai-as-promised';
-import notFound from '../../src/notfound/handler.js';
+import noopHandler from '../../src/digest/handler-noop.js';
 
+chai.use(sinonChai);
 chai.use(chaiAsPromised);
 const { expect } = chai;
 
-describe('not found handler', () => {
+const sandbox = sinon.createSandbox();
+
+describe('noop handler', () => {
   let message;
   let context;
+  let mockLog;
 
   beforeEach('setup', () => {
     message = {
       url: 'space.cat',
+      type: 'some-type',
+    };
+    mockLog = {
+      info: sinon.spy(),
+      warn: sinon.spy(),
+      error: sinon.spy(),
     };
     context = {
-      log: console,
-      env: {
-        SLACK_BOT_TOKEN: 'token',
-        RUM_DOMAIN_KEY: 'uber-key',
-      },
+      log: mockLog,
     };
   });
 
-  it('executes noop when receiving a 404 report request', async () => {
-    const resp = await notFound(message, context);
+  afterEach('clean', () => {
+    sandbox.restore();
+  });
+
+  it('executes noop and just logs alert report type and url', async () => {
+    const resp = await noopHandler(message, context);
+
     expect(resp.status).to.equal(204);
+
+    expect(mockLog.info).to.have.been.calledWith(`Alert report of type ${message.type} received for: ${message.url}`);
   });
 });
